Made FFT in place

+++ include/Datatypes.h
@@ -43,7 +43,7 @@ typedef float4 Real4;
 typedef float2 Complex;
 
 #endif
-typedef uint64_t BigUINT;
+typedef uint32_t BigUINT;
 typedef uint32_t UINT;
 
 #define NUM_THREADS 128
